Condições

+++ Apps/Google Chrome/reinstall_chrome.sh
@@ -2,10 +2,20 @@
 
 ## Created by Gabriel Barbato to reinstall Chrome - https://github.com/barbatoo
 
-sudo killall Google\ Chrome
+if pgrep "Google Chrome" > /dev/null; then
+    sudo killall "Google Chrome"
+    echo "Processo do Google Chrome encerrado, inciciando desisntalação..."
+else
+    echo "Não há processo do Google Chrome aberto, inciando verificação do Chrome..."
+fi
 
-rm -rf "/Applications/Google Chrome.app"
+if [ -d "/Applications/Google Chrome.app" ] 
+then
+    rm -rf "/Applications/Google Chrome.app"
     echo "Google Chrome desinstalado, iniciando reinstalação..."
+else
+    echo "Google Chrome não está instalado, inciando instalação..."
+fi
 
 curl -o /tmp/googlechrome.dmg "https://dl.google.com/chrome/mac/universal/stable/GGRO/googlechrome.dmg"
 
@@ -25,6 +35,6 @@ then
     echo "O Google Chrome foi atualizado com sucesso!"
     exit 0
 else
-    echo "O Google Chrome não pode ser atualizado."
+    echo "O Google Chrome não pode ser atualizado/instalado."
     exit 1
 fi
